% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_archive.R
\name{archive_connect}
\alias{archive_connect}
\title{Archive Connect}
\usage{
archive_connect(output_dir, commit, fl)
}
\arguments{
\item{output_dir}{the directory where the archive structure is}

\item{commit}{the commit id}

\item{fl}{the file name to extract}
}
\value{
the archive connection
}
\description{
Extracts a csv (comma delimited) file from the archive
}
