% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elt.R
\name{archive_elt}
\alias{archive_elt}
\title{Initiates the ETL for archiving}
\usage{
archive_elt(output_dir, fls)
}
\arguments{
\item{output_dir}{the output directory}

\item{fls:}{the files to move to Current folder}
}
\description{
Initiates the ETL for archiving
}
