% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{archive_etl}
\alias{archive_etl}
\title{Archive ETL}
\usage{
archive_etl(output_dir, fls)
}
\arguments{
\item{output_dir}{the output directory}

\item{fls:}{the files to move to Current folder}
}
\description{
Initiates the ETL for archiving
}
