% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive_output}
\alias{archive_output}
\title{Writes archive fls out to the Current folder and archives the same files with a commit id}
\usage{
archive_output(commit, output_dir, fls)
}
\arguments{
\item{commit}{the commit id}

\item{output_dir}{the path to create the file archive}

\item{fls}{a list of the files to archive}
}
\description{
Writes archive fls out to the Current folder and archives the same files with a commit id
}
\examples{
commit <- 'master_98jgh8'
output_dir <- '.'
fls <- c('data.csv', 'data_log.log')
archive_output(commit, output_dir, fls)
}
