% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitTools.R
\name{auto_commit}
\alias{auto_commit}
\title{Automattically commits}
\usage{
auto_commit(repo, user.name, user.password, fl, message, add_branch,
  add_message)
}
\arguments{
\item{repo}{the repository}

\item{user.name}{username}

\item{user.password}{user password}

\item{fl}{the file to commit}

\item{message}{the commit message}

\item{add_branch}{takes the values TRUE or FALSE; if TRUE adds branch to the beginning of the the commit tag}

\item{add_message}{takes the values TRUE or FALSE; if TRUE adds message to the end of the commit tag}
}
\value{
filename to be used to version output files and datasets
}
\description{
Automattically commits code
}
\examples{
repo <- config.repo('Joli Holmes', 'holmesjoligmail.com')
auto_commit(repo, 'holmesjoli', 'password', '.', 'July 2018 Monthly')
}
