% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitTools.R
\name{check_master}
\alias{check_master}
\title{Checks master branch active}
\usage{
check_master(repo)
}
\arguments{
\item{repo}{the repository to check}
}
\value{
the name of the active branch
}
\description{
Checks to see if the HEAD is on the master branch
}
\examples{
repo <- config.repo('Joli Holmes', 'holmesjoligmail.com')
check_master(repo)
}
