% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{config_repo}
\alias{config_repo}
\title{Configure repository}
\usage{
config_repo(user.name, user.email)
}
\arguments{
\item{user.name}{server (github, bitbucket, etc.) username}

\item{user.email}{the users email}
}
\value{
the repository
}
\description{
Initiates and configures the repository
}
\examples{
repo <- config.repo("Joli Holmes", "holmesjoli\\@\\gmail.com")
}
