% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{create_archive}
\alias{create_archive}
\title{Create archive}
\usage{
create_archive(commit, current_dir, archive_dir)
}
\arguments{
\item{commit}{the commit id}

\item{current_dir}{the current directory}

\item{archive_dir}{the archive directory}
}
\description{
Copies files from the Current folder, adds them to the archive, and zips files using the commit id as the filename
}
\examples{
commit <- 'master_98jgh8'
output_dir <- '.'
archive_files <- c('data.csv', 'data_log.log')
create_archive(commit, output_dir, fls)
}
