% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{create_archive_dir}
\alias{create_archive_dir}
\title{Creates the file path to the Archive directory}
\usage{
create_archive_dir(output_dir)
}
\arguments{
\item{output_dir}{the path to create the versioning folder structure in}
}
\description{
Creates the file path to the Archive directory
}
\examples{
ad <- archive_dir('C:/Projects/ProjectXYZ/Data/Processed')
}
