% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{create_archive_str}
\alias{create_archive_str}
\title{Creates the archive structure}
\usage{
create_archive_str(output_dir)
}
\arguments{
\item{output_dir}{the output directory where the archive structure will be created}
}
\description{
Creates the archive structure
}
\examples{
output_dir <- c('./Data/Processed')
create_archive_str(output_dir)
}
