% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{create_current_dir}
\alias{create_current_dir}
\title{Creates the file path to the Current directory}
\usage{
create_current_dir(output_dir)
}
\arguments{
\item{output_dir}{the path to create the versioning folder structure in}
}
\description{
Creates the file path to the Current directory
}
\examples{
cd <- current_dir('C:/Projects/ProjectXYZ/Data/Processed')
}
