% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_archive.R
\name{extract_csv}
\alias{extract_csv}
\title{Extracts a csv (comma delimited) file from the archive}
\usage{
extract_csv()
}
\arguments{
\item{output_dir}{the directory where the archive structure is}

\item{commit}{the commit id}

\item{fl}{the file name to extract}
}
\value{
the data from the file specified
}
\description{
Extracts a csv (comma delimited) file from the archive
}
\examples{
output_dir <- './tests/testthat/Data'
commit <- 'bd429f' #this will have to be changed
fl <- 'data.csv'
df <- extract_csv(output_dir, commit, fl)
}
