% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{move_to_current}
\alias{move_to_current}
\title{Moves the files to the Current folder}
\usage{
move_to_current(archive_fls, current_dir)
}
\arguments{
\item{archive_fls}{a list of the files to archive}

\item{current_dir}{the current directory}
}
\description{
Moves the files to the Current folder
}
\examples{
output_dir <- '.'
current_dir <- create_current_dir(output_dir)
archive_fls <- c('data.csv', 'data_log.log')
utilsR::create_files(archive_fls)
move_to_current(archive_fls, current_dir)
}
