% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_args.R
\name{parseArguments}
\alias{parseArguments}
\title{Parse arguments}
\usage{
parseArguments()
}
\description{
Parses different options from the command line
}
\examples{
\dontrun{
R/sample_code.R -c "commit message" -e holmesjoli@gmail.com -n holmesjoli -p gitHubPassword`
parseArguments()
}
}
