% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{pass_query}
\alias{pass_query}
\title{Wrapper for web.archive.org/save}
\usage{
pass_query(query_url, cookies)
}
\arguments{
\item{query_url}{A url string of a page that is to be archived}

\item{cookies}{A cookie string created by prepare_auth()}
}
\description{
This function builds a command to save a specified URL to archive.org.
NB: it is recommended to have cookies saved as a string by running prepare_auth()
}
\examples{
pass_query(query_url = "https://cnn.com", cookies = cookies)
}
\keyword{archive,}
\keyword{waybackmachine}
