% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{prepare_auth}
\alias{prepare_auth}
\title{Prepare authorization cookie string}
\usage{
prepare_auth()
}
\description{
This function is step by step process to create an auth cookie string for archive.org
Recommended to run this once and save the output as a named string. NB. Cookies expire, so this should be repeated regularly.
}
\examples{
cookies <- prepare_auth()
}
\keyword{archive,}
\keyword{waybackmachine}
