% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitTools.R
\name{setUp}
\alias{setUp}
\title{Configure repository}
\usage{
setUp(user.name, user.email)
}
\arguments{
\item{user.name}{server (github, bitbucket, etc.) username}

\item{user.email}{the users email}
}
\value{
the repository
}
\description{
Initiates and configures the repository
}
\examples{
repo <- config.repo('Joli Holmes', 'holmesjoligmail.com')
}
