% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{web_archive_list}
\alias{web_archive_list}
\title{Save a list of urls to web.archive.org}
\usage{
web_archive_list(url_list, cookies)
}
\arguments{
\item{url_list}{A list of url strings}

\item{cookies}{A cookie string created by prepare_auth()}
}
\description{
This function interates through a list of url strings to save to archive.org.
NB: it is recommended to have cookies saved as a string by running prepare_auth()
}
\keyword{archive,}
\keyword{waybackmachine}
