% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitTools.R
\name{auto_commit}
\alias{auto_commit}
\title{Automattically commits}
\usage{
auto_commit(repo, user.name, user.password, fl, message)
}
\arguments{
\item{repo}{the repository}

\item{user.name}{username}

\item{user.password}{user password}

\item{fl}{the file to commit}

\item{message}{the commit message}
}
\value{
filename to be used to version output files and datasets
}
\description{
Automattically commits code
}
\examples{
repo <- config.repo('Joli Holmes', 'holmesjoligmail.com')
auto_commit(repo, 'holmesjoli', 'password', '.', 'July 2018 Monthly')
}
